requirements.txt
alembic==1.5.4
appdata==2.1.2
attrs==21.2.0
autopep8==1.5.5
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colored==1.4.3
coverage==6.3.2
distlib==0.3.4
filelock==3.6.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.22.3
packaging==20.9
pbr==5.8.1
platformdirs==2.5.1
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
stevedore==3.5.0
to-requirements.txt==1.3.0
toml==0.10.2
urllib3==1.26.4
virtualenv==20.13.2
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
Werkzeug==1.0.1
zipp==3.8.0
